fix: resolve all 34 Pylance errors across 10 files
- report_generator.py: Add datetime imports
- resilience.py: Fix device_id_b parameter, Optional return types
- metrics.py: Fix max() calls with None filtering on sync times
- performance.py: Add Optional[dict] return type
- dsp.py: Move delay_samples declaration before use
- integration_test.py: Add missing round_num params, fix parameter names
- tournament.py: Fix float vs int type hint for fastest_sync_ms
- audio.py: Add missing device, format, and p attributes to AudioHandler
- self_test.py: Pass config dict to AudioHandler instantiations
- dashboard/__init__.py: Create missing module init file
- nightly-chaos.yml: Add condition check for SLACK_WEBHOOK_URL secret

All type errors resolved, code now passes Pylance strict mode.

+++ raspberry_pi_server/tournament.py
@@ -3,14 +3,14 @@ Tournament mode: support for multi-station (up to 8) competitive bracket.
 Handles pairing, tournament progression, leaderboard.
 """
 
+import json
+import logging
+import random
 import threading
 import time
-import random
-import logging
-from enum import Enum
-from typing import Dict, List, Tuple, Optional
 from dataclasses import dataclass, field
-import json
+from enum import Enum
+from typing import Dict, List, Optional, Tuple
 
 logger = logging.getLogger(__name__)
 
@@ -45,7 +45,7 @@ class Leaderboard:
     station_id: str
     matches_played: int = 0
     matches_won: int = 0
-    fastest_sync_ms: int = float('inf')
+    fastest_sync_ms: float = float('inf')
     avg_sync_ms: float = 0.0
     total_sync_ms: int = 0
